requirements.txt
# Core
fastapi==0.128.2
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.7.1

# Database
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.46
alembic==1.18.3

# Scraping
httpx==0.28.1
beautifulsoup4==4.12.3

# Cache
redis==5.0.1

# Payments
stripe==8.0.0

# Utils
python-dotenv==1.0.0
jinja2==3.1.6
python-multipart==0.0.22

# Security
cryptography==46.0.4

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
